Moved other settings to its own class
