+++ Assets/Scripts/Settings/UISettingsOther.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 49e773074de12594e9980542be5a673b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
